Correcting the component-ization. paul
