+++ api/src/main/resources/META-INF/components.txt
@@ -1,1 +1,2 @@
 org.xwiki.apps.googleapps.GoogleAppsScriptService
+org.xwiki.apps.googleapps.internal.CookieAuthenticationPersistenceImpl
